requirements.txt
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
easyocr==1.4.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.2.0/en_core_web_lg-3.2.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
entrypoints==0.4
executing==0.8.3
idna==3.3
imageio==2.16.2
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.2.2
jupyter-core==4.9.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nest-asyncio==1.5.5
networkx==2.8
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkg_resources==0.0.0
pluggy==1.0.0
preshed==3.0.6
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.8
pytesseract==0.3.9
pytest==7.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data==0.2.0
thinc==8.0.15
tifffile==2022.4.8
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
XlsxWriter==3.0.3
